// src/types/character-book.ts
export type CharacterBook = {
    name?: string;
    description?: string;
    scan_depth?: number;
    token_budget?: number;
    recursive_scanning?: boolean;
    extensions: Record<string, any>;
    entries: Array<CharacterBookEntry>;
}

export type CharacterBookEntry = {
    keys: Array<string>;
    content: string;
    extensions: Record<string, any>;
    enabled: boolean;
    insertion_order: number;
    case_sensitive?: boolean;

    name?: string;
    priority?: number;

    id?: number;
    comment?: string;
    selective?: boolean;
    secondary_keys?: Array<string>;
    constant?: boolean;
    position?: 'before_char' | 'after_char';
};